refactor: restructure stream
