+++ duva/src/presentation/clients/controller.rs
@@ -1,13 +1,34 @@
-use crate::{
-    domains::cluster_actors::commands::ConsensusClientResponse,
-    presentation::clients::request::ClientRequest,
-    presentation::clusters::connection_manager::ClusterConnectionManager,
-};
+use super::request::ClientRequest;
+use crate::actor_registry::ActorRegistry;
+use crate::domains::caches::cache_manager::CacheManager;
+use crate::domains::caches::cache_objects::CacheEntry;
+use crate::domains::cluster_actors::commands::{ClusterCommand, ConsensusClientResponse};
+use crate::domains::config_actors::command::ConfigResponse;
+use crate::domains::config_actors::config_manager::ConfigManager;
+use crate::domains::query_parsers::QueryIO;
+use crate::domains::saves::actor::SaveTarget;
+use crate::presentation::clients::request::ClientAction;
+use crate::presentation::clusters::communication_manager::ClusterCommunicationManager;
+use crate::presentation::clusters::connection_manager::ClusterConnectionManager;
+use futures::future::try_join_all;
 use std::sync::atomic::Ordering;
 
-use super::*;
+#[derive(Clone)]
+pub(crate) struct ClientController {
+    pub(crate) cache_manager: CacheManager,
+    pub(crate) config_manager: ConfigManager,
+    pub(crate) cluster_communication_manager: ClusterCommunicationManager,
+}
 
 impl ClientController {
+    pub(crate) fn new(actor_registry: ActorRegistry) -> Self {
+        Self {
+            cluster_communication_manager: actor_registry.cluster_communication_manager(),
+            cache_manager: actor_registry.cache_manager,
+            config_manager: actor_registry.config_manager,
+        }
+    }
+
     pub(crate) async fn handle(
         &self,
         cmd: ClientAction,
